moving stuff to source files
