+++ include/prismspf/core/pde_operator.h
@@ -79,26 +79,4 @@ private:
   const userInputParameters<dim> *user_inputs;
 };
 
-template <int dim, int degree, typename number>
-PDEOperator<dim, degree, number>::PDEOperator(
-  const userInputParameters<dim> &_user_inputs)
-  : user_inputs(&_user_inputs)
-{}
-
-template <int dim, int degree, typename number>
-inline const userInputParameters<dim> &
-PDEOperator<dim, degree, number>::get_user_inputs() const
-{
-  Assert(user_inputs != nullptr, dealii::ExcNotInitialized());
-  return *user_inputs;
-}
-
-template <int dim, int degree, typename number>
-inline number
-PDEOperator<dim, degree, number>::get_timestep() const
-{
-  Assert(user_inputs != nullptr, dealii::ExcNotInitialized());
-  return user_inputs->temporal_discretization.dt;
-}
-
 PRISMS_PF_END_NAMESPACE
