+++ src/solvers/explicit_constant_solver.cc
@@ -1,0 +1,49 @@
+#include <prismspf/solvers/explicit_constant_solver.h>
+
+#include <prismspf/config.h>
+
+PRISMS_PF_BEGIN_NAMESPACE
+
+template <int dim, int degree>
+explicitConstantSolver<dim, degree>::explicitConstantSolver(
+  const userInputParameters<dim>                         &_user_inputs,
+  const matrixfreeHandler<dim>                           &_matrix_free_handler,
+  const invmHandler<dim, degree>                         &_invm_handler,
+  const constraintHandler<dim>                           &_constraint_handler,
+  const dofHandler<dim>                                  &_dof_handler,
+  const dealii::MappingQ1<dim>                           &_mapping,
+  solutionHandler<dim>                                   &_solution_handler,
+  std::shared_ptr<const PDEOperator<dim, degree, double>> _pde_operator)
+  : explicitBase<dim, degree>(_user_inputs,
+                              _matrix_free_handler,
+                              _invm_handler,
+                              _constraint_handler,
+                              _dof_handler,
+                              _mapping,
+                              _solution_handler,
+                              _pde_operator)
+{}
+
+template <int dim, int degree>
+void
+explicitConstantSolver<dim, degree>::init()
+{
+  this->compute_subset_attributes(fieldSolveType::EXPLICIT_CONSTANT);
+
+  // If the subset attribute is empty return early
+  if (this->subset_attributes.empty())
+    {
+      return;
+    }
+
+  this->set_initial_condition();
+}
+
+template <int dim, int degree>
+void
+explicitConstantSolver<dim, degree>::solve()
+{}
+
+INSTANTIATE_BI_TEMPLATE(explicitConstantSolver)
+
+PRISMS_PF_END_NAMESPACE
