+++ src/solvers/nonexplicit_auxiliary_solver.cc
@@ -1,0 +1,121 @@
+#include <prismspf/solvers/nonexplicit_auxiliary_solver.h>
+
+#include <prismspf/config.h>
+
+PRISMS_PF_BEGIN_NAMESPACE
+
+template <int dim, int degree>
+nonexplicitAuxiliarySolver<dim, degree>::nonexplicitAuxiliarySolver(
+  const userInputParameters<dim>                         &_user_inputs,
+  const matrixfreeHandler<dim>                           &_matrix_free_handler,
+  const triangulationHandler<dim>                        &_triangulation_handler,
+  const invmHandler<dim, degree>                         &_invm_handler,
+  const constraintHandler<dim>                           &_constraint_handler,
+  const dofHandler<dim>                                  &_dof_handler,
+  const dealii::MappingQ1<dim>                           &_mapping,
+  dealii::MGLevelObject<matrixfreeHandler<dim, float>>   &_mg_matrix_free_handler,
+  solutionHandler<dim>                                   &_solution_handler,
+  std::shared_ptr<const PDEOperator<dim, degree, double>> _pde_operator)
+  : nonexplicitBase<dim, degree>(_user_inputs,
+                                 _matrix_free_handler,
+                                 _triangulation_handler,
+                                 _invm_handler,
+                                 _constraint_handler,
+                                 _dof_handler,
+                                 _mapping,
+                                 _mg_matrix_free_handler,
+                                 _solution_handler,
+                                 _pde_operator)
+{}
+
+template <int dim, int degree>
+inline void
+nonexplicitAuxiliarySolver<dim, degree>::init()
+{
+  this->compute_subset_attributes(fieldSolveType::NONEXPLICIT_AUXILIARY);
+
+  // If the subset attribute is empty return early
+  if (this->subset_attributes.empty())
+    {
+      return;
+    }
+
+  for (const auto &[index, variable] : this->subset_attributes)
+    {
+      // Creating temporary map to match types
+      std::map<unsigned int, variableAttributes> temp;
+      temp.emplace(index, variable);
+      subset_attributes_list.push_back(temp);
+
+      // Create the implementation of matrixFreeOperator with the subset of variable
+      // attributes
+      this->system_matrix[index] =
+        std::make_unique<SystemMatrixType>(subset_attributes_list.back(),
+                                           this->pde_operator,
+                                           index);
+
+      // Set up the user-implemented equations and create the residual vectors
+      this->system_matrix.at(index)->clear();
+      this->system_matrix.at(index)->initialize(
+        this->matrix_free_handler->get_matrix_free());
+
+      // Create the subset of solution vectors and add the mapping to matrixFreeOperator
+      new_solution_subset[index].push_back(
+        this->solution_handler->get_new_solution_vector(index));
+      solution_subset[index].push_back(
+        this->solution_handler->get_solution_vector(index, dependencyType::NORMAL));
+      global_to_local_solution[index].emplace(std::make_pair(index,
+                                                             dependencyType::NORMAL),
+                                              0);
+      for (const auto &[variable_index, map] :
+           subset_attributes_list.back().begin()->second.dependency_set_RHS)
+        {
+          for (const auto &[dependency_type, field_type] : map)
+            {
+              const auto pair = std::make_pair(variable_index, dependency_type);
+
+              solution_subset[index].push_back(
+                this->solution_handler->get_solution_vector(variable_index,
+                                                            dependency_type));
+              global_to_local_solution[index].emplace(pair,
+                                                      solution_subset.at(index).size() -
+                                                        1);
+            }
+        }
+      this->system_matrix.at(index)->add_global_to_local_mapping(
+        global_to_local_solution.at(index));
+    }
+}
+
+template <int dim, int degree>
+inline void
+nonexplicitAuxiliarySolver<dim, degree>::solve()
+{
+  // If the subset attribute is empty return early
+  if (this->subset_attributes.empty())
+    {
+      return;
+    }
+
+  for (const auto &[index, variable] : this->subset_attributes)
+    {
+      // Compute the update
+      this->system_matrix.at(index)->compute_nonexplicit_auxiliary_update(
+        new_solution_subset.at(index),
+        solution_subset.at(index));
+
+      // Scale the update by the respective (SCALAR/VECTOR) invm.
+      new_solution_subset.at(index).at(0)->scale(this->invm_handler->get_invm(index));
+
+      // Update the solutions
+      this->solution_handler->update(fieldSolveType::NONEXPLICIT_AUXILIARY, index);
+
+      // Apply constraints
+      this->constraint_handler->get_constraint(index).distribute(
+        *(this->solution_handler->get_solution_vector(index, dependencyType::NORMAL)));
+    }
+}
+
+INSTANTIATE_BI_TEMPLATE(nonexplicitAuxiliarySolver)
+
+PRISMS_PF_END_NAMESPACE
