+++ src/utilities/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Manually specify files to be included
+list(APPEND PRISMS_PF_SOURCE_FILES
+  ${CMAKE_CURRENT_SOURCE_DIR}/compute_integral.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/element_volume.cc
+)
+set(PRISMS_PF_SOURCE_FILES ${PRISMS_PF_SOURCE_FILES} PARENT_SCOPE)
